tests/wpt/mozilla/tests/bluetooth/getAvailability/adapter-off.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.not_powered);
    return window.navigator.bluetooth.getAvailability()
    .then(isAvailable => assert_equals(isAvailable, false));
}, 'GetAvailability is false if the adapter is off.');
</script>
